+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to the "vscode-camelk" extension will be documented in this 
 
 ## 0.0.25
 
-- Provide specific `camel-k-debug`  VS Code tasks
+- Debugging:
+  - Provide right-click menu in Integration view to attach a Java debugger. It requires to be launched on localhost.
+  - Provide specific `camel-k-debug`  VS Code tasks
 
 ## 0.0.24